Recalculated insets bounding box, now markers work
Initially the bounding box for the insets was the same as the lat/long range.
This is not how the `jquery-jvectormap-world-mill-en.js` defines the insets.

Working backwards from the world-mill-en map, I found the bbox was 44,478.7439403058 times the pixel size, so I applied this same constant to the bbox for the tdwg inset, and got the following:

```json
[
  {
    "bbox": [
      {
        "x": -22239371.9701529,
        "y": -10331868.117680805
      },
      {
        "x": 22239371.9701529,
        "y": 11119685.98507645
      }
    ],
    "height": 482.28776719835065,
    "left": 0,
    "top": 0,
    "width": 1000.0
  }
]
```