Fix reset button
Reset was broken because the 4 individual search controllers maintained
a reference to their part of the SearchModel object. When the
SearchCtrl reset the SearchModel object, the controllers still had
their reference to the stale object.

Solved with a $scope.$broadcast and $scope.$on
Considered a $scope.$watch / $watchCollection, but the whole point of
the searchModelService is to be a singleton and not live in a
controller scope. May be in the root scope, but it seemed more explicit
to broadcast an event.