Upgrade dependencies
To make sure dplace can run with recent versions of its dependencies, the following
changes have been necessary:

- Switch from migrations via South to Django's own system.
- Adapt the settings templates to add a `TEST_RUNNER` setting.

Upgrading to django-rest-framework 3.3.2 involved small changes in the code. It must
be noted, though, that the format of JSON responses seems to have changed. The
necessary adaptions in the tests have been made, but it is clear, that the corresponding
changes of the frontend code must be implemented.