Refactored load system
This turned out to be a rather big rewrite of the system for data loading.
The goal was to make data loading faster, and on my machine it went from
about 55min to 6.5min - which makes a **big** difference. In particular, it
is now practical to always recreate the database and load a complete set of
data, which should make the data loading process more reproducible, since no
broken remains of older data loads will linger in the database.