Convert find_societies API call to GET request
First naive implementation of find_societies as GET request. The length of
the search criteria passed to the server is still quite big, also due to the
overhead of the URL-escaped JSON syntax elements in the query.
E.g. searching for 2 regions and 163 languages results in a URL of length ~2700.

Also closes #271