Optimized society serializer for language trees
This commit exemplifies how serializers could be optimized for better
performance. The data about the language and source a society is related
to is already present in the `societies` member of a search results object,
thus it doesn't have to be serialized again for the society objects related
to taxa in a `LanguageTree`. This optimization alone cuts the time for a
"tree-intensive" search like "culture->EA->Class->EA066" on my machine from
about 10 secs to less than 6 secs.