Make datasets to be loaded configurable
In addition to adapting the load process to the new file names introduced in
the data repository, the list of datasets to be loaded is now a setting in
`dplace/settings.py`, which can be accessed as `django.conf.settings.DATASETS`.

This makes for less code in the load scripts as well as for easier addition of
more datasets.

Note that deploying this change requires
- updating the local `dplace/settings.py`
- updating the data repos clone
- re-loading the data