Revert back since issue is a front-end problem
When switching to the Results tab the front-end angular code issues
a second search request (in the case of society-by-name search).
This shouldn't happen. The issue at hand was due to the front-end
sending the second query encoded differently. This has been fixed
with this commit, but the bigger issue still is unsolved.